+++ app/src/pages/MethodsPage.test.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+
+import { shallow, ShallowWrapper } from "enzyme";
+import { MethodsPage } from "./MethodsPage";
+
+describe('<MethodsPage>', () => {
+    let wrapper: ShallowWrapper;
+
+    beforeEach(() => {
+        wrapper = shallow(<MethodsPage/>);
+    });
+
+    it('should render', () => {
+        expect(wrapper.isEmptyRender()).toBeFalsy();
+    })
+});