+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Column check for uploading of genome links ([#154](https://github.com/iomega/paired-data-form/pull/154))
+* Methods page
 
 ### Fixed