+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2020-09-22
+
 ### Added
 
 * Column check for uploading of genome links ([#154](https://github.com/iomega/paired-data-form/pull/154))
@@ -16,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Render publications delimited by spaces ([#152](https://github.com/iomega/paired-data-form/pull/152))
 
-## [0.7.0] - 2020-10-07
+## [0.7.0] - 2020-07-10
 
 ### Added